+++ lib/fixture_champagne/migrator.rb
@@ -22,10 +22,6 @@ module FixtureChampagne
       def tmp_fixture_path
         Rails.root.join("tmp", "fixtures")
       end
-
-      def fixture_attachment_folders
-        %w[files active_storage action_text].map { |f| single_fixture_path.join(f) }
-      end
     end
 
     def initialize(
@@ -168,9 +164,13 @@ module FixtureChampagne
     end
 
     def copy_fixture_attachments
-      self.class.fixture_attachment_folders.each do |folder|
+      fixture_attachment_folders.each do |folder|
         FileUtils.cp_r(folder, self.class.tmp_fixture_path) if folder.exist?
       end
+    end
+
+    def fixture_attachment_folders
+      %w[files active_storage action_text].map { |f| effective_fixture_path.join(f) }
     end
 
     def temporary_fixture_filename(klass)
@@ -188,10 +188,33 @@ module FixtureChampagne
     end
 
     def overwrite_fixtures
-      removable_fixture_path = self.class.single_fixture_path.dirname.join("old_fixtures")
-      FileUtils.mv(self.class.single_fixture_path, removable_fixture_path)
-      FileUtils.mv(self.class.tmp_fixture_path, self.class.single_fixture_path)
+      removable_fixture_path = effective_fixture_path.dirname.join("old_fixtures")
+      FileUtils.mv(effective_fixture_path, removable_fixture_path)
+      FileUtils.mv(tmp_fixture_path, effective_fixture_path)
       FileUtils.rm_r(removable_fixture_path, secure: true)
+    end
+
+    def effective_fixture_path
+      @effective_fixture_path ||= set_effective_fixture_path
+    end
+
+    def set_effective_fixture_path
+      # From Rails 7.1 users can set multiple fixture paths, the default being test/fixtures.
+      # Most users will use the default path. Others may use a custom path but keep the default in the array,
+      # even if they do not use it. For now the gem supports only these 2 cases.
+      # If a user has fixture files in multiple paths, then it's trickier to decide where to save the new
+      # generated fixtures, so for now the gem will not allow these users to overwrite current files.
+
+      paths_with_files = MigrationContext.fixture_paths.select do |path|
+        Dir[::File.join(path, "{**,*}/*.{yml}")].any?
+      end
+
+      if paths_with_files.size > 1 && configuration.overwrite_fixtures?
+        raise OverwriteNotAllowedError,
+              "can't overwrite fixtures when using multiple folders: #{path_with_files.to_json}. Set overwrite to false"
+      end
+
+      paths_with_files.first
     end
 
     def remember_new_fixture_versions