+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - Add support for jsonb and encrypted attributes
 - Use `TestFixtures#fixture_paths` intead of the deprecated method in singular `TestFixture#fixture_path`
+- Add partial support for multiple fixture paths
 
 ## [0.1.0] - 2023-01-23